backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
ddgs>=9.13.0
googlesearch-python>=1.3.0
pytest>=8.0.0
httpx>=0.27.0
